+++ frontend/src/layout/lemonade/SideBar/SideBar.scss
